// Frontend/src/app/model/x_auftrag_flt.ts
export class XAuftragFlt {
    name: string;
    GUID: string;
    SO_ID: string;
    EO_ID: string;
    ORDER_ID_CRM: string;
    ORDER_ID_BSSOE: string;
    ORDER_ID_SMF: string;
    ORDER_ID_DKK: string;
    ORDER_ID_SNR: string;
    ORDER_ID_MISC: string;
    constructor() {
        console.log("class XAuftragFlt");
        this.name='XAuftragFlt';
        this.GUID='';
        this.SO_ID='';
        this.EO_ID='';
        this.ORDER_ID_CRM='';
        this.ORDER_ID_BSSOE='';
        this.ORDER_ID_SMF='';
        this.ORDER_ID_DKK='';
        this.ORDER_ID_SNR='';
        this.ORDER_ID_MISC='';
        }
}